Revert "feat: creates wallet settings screen"
This reverts commit ee670a2359993ca7434a0fdf2012485c4cd42ff2.
