+++ lib/common/utils/colors.dart
@@ -18,5 +18,3 @@ Color get color24252C => Color(0xFF24252C);
 Color get colorD64CA8 => Color(0xFFD64CA8);
 Color get color1F1F1F => Color(0xFF1F1F1F);
 Color get color93949A => Color(0xFF93949A);
-Color get color8E8E93 => Color(0xFF8E8E93);
-Color get color02C076 => Color(0xFF02C076);
